refactor: rename screen files
