+++ packages/app/navigation/index.tsx
@@ -18,16 +18,16 @@ import { useColorScheme, useServer } from "../hooks/use-settings";
 import  { useTheme } from "../hooks/use-theme-color";
 import { RootStackParamList, SettingsStackParamList } from "../types";
 
-import NotFoundScreen from "../screens/not-found-screen";
-import TorrentsScreen from "../screens/torrents-screen";
-import SettingsScreen from "../screens/settings-screen";
-import AboutScreen from "../screens/about-screen";
-import AddTorrentMagnetScreen from "../screens/add-torrent-magnet-screen";
-import AddTorrentFileScreen from "../screens/add-torrent-file-screen";
-import ConnectionSetupScreen from "../screens/connection-setup-screen";
-import ServerConfigurationScreen from "../screens/server-configuration-screen";
-import ThemeScreen from "../screens/theme-screen";
-import TorrentDetails from "../screens/torrent-details";
+import NotFoundScreen from "../screens/not-found";
+import TorrentsScreen from "../screens/torrents";
+import SettingsScreen from "../screens/settings";
+import AboutScreen from "../screens/about";
+import AddTorrentMagnetScreen from "../screens/add-torrent-magnet";
+import AddTorrentFileScreen from "../screens/add-torrent-file";
+import ConnectionSetupScreen from "../screens/connection-setup";
+import ServerConfigurationScreen from "../screens/server-configuration";
+import ThemeScreen from "../screens/theme";
+import TorrentDetailsScreen from "../screens/torrent-details";
 
 export function NavigationContainer({
   children,
@@ -73,7 +73,7 @@ export function RootNavigator() {
           <>
             <Stack.Screen
               name="TorrentDetails"
-              component={TorrentDetails}
+              component={TorrentDetailsScreen}
               options={{ title: "Details" }}
             />
